add ops
