+++ quantization/ops/__init__.py
@@ -1,0 +1,4 @@
+from conv import QConv2d as q_conv
+from linear import QLinear as q_linear
+from pool import QMaxPooling2d as q_max_pool_2d
+from relu import QReLU as q_relu
