+++ quantization/ops/relu.py
@@ -1,0 +1,31 @@
+import oneflow as flow
+from ..q_module import QParam, QModule
+
+class QReLU(QModule):
+
+    def __init__(self, qi=False):
+        super(QReLU, self).__init__(qi=qi)
+
+    def freeze(self, qi=None):
+        
+        if hasattr(self, 'qi') and qi is not None:
+            raise ValueError('qi has been provided in init function.')
+        if not hasattr(self, 'qi') and qi is None:
+            raise ValueError('qi is not existed, should be provided.')
+
+        if qi is not None:
+            self.qi = qi
+
+    def forward(self, x):
+        if hasattr(self, 'qi'):
+            self.qi.update(x)
+            x = self.qi.fake_quantize_tensor(x)
+
+        x = flow.F.relu(x)
+
+        return x
+    
+    def quantize_inference(self, x):
+        x = x.clone()
+        x[x < self.qi.zero_point] = self.qi.zero_point
+        return x
